Move `TCPServer` to `netutil`, change `handle_stream` callback to overridden method hook, move IP address hack for unix domain sockets to `HTTPConnection`
